feat: enable multi-image per sample support in emu3 and emu3p5 encoder models. Properly load emu3p5 vision tokenizer in emu3p5 encoder based models.
